Add unsafe `renderUnsafe()` and `getUnsafeRenderer()`
This gives the developer an option to use a side-effecty render function
that throws an exception.

The use case for this is when we are not able to render an error page.
If something goes wrong then, it's probably better to just fail fast,
then getting stuck in a loop trying to render more error-pages.

Closes: #25